refactor: improve state management and popover handling in HeaderLeftToolBarInput and useSearchPopover
- Replaced local state for popover visibility in HeaderLeftToolBarInput with a centralized state from useSearchPopover, enhancing consistency across components.
- Introduced handleInputBlur to manage popover visibility on input blur, improving user experience.
- Updated related components to utilize the new state management logic, streamlining interactions and reducing redundant code.
